+++ internal/testutils/netnsutils/context.go
@@ -1,4 +1,4 @@
-package testutils
+package netnsutils
 
 import "testing"
