+++ internal/testutils/netnsutils/context_windows.go
@@ -1,4 +1,4 @@
-package testutils
+package netnsutils
 
 import "testing"
