+++ internal/testutils/netnsutils/sanity_linux.go
@@ -1,4 +1,4 @@
-package testutils
+package netnsutils
 
 import (
 	"errors"
